ub: udma: Support loading and unloading driver.
commit 202d693228557be2d91043c69023586a372cc876 openEuler.

This patch adds the function of loading and unloading the driver.
In driver loading process, udma queries the required hardware
information from ubase and stores it in the internal structure of the
driver. At the same time, udma also provides ubcore with ops for
querying device status and create rc table. This patch also adds the
function of mailbox and software table for udma driver.

Signed-off-by: Wei Qin <qinwei61@huawei.com>
Signed-off-by: zhaoweibo <zhaoweibo3@huawei.com>
Signed-off-by: huwentao <huwentao19@h-partners.com>