obmm: set MAX Numa Distance to 255
ANCK ID: #IJSBBS

commit 83ae17336e75ca20f3819afd45e2067346ca6f77 openEuler.

We used to considered 255 as a reserved value for numa_distance.
Now we have confirmed that 255 indicates a distance which is unreachable.
We adhere to this semantics and set this value as the maximum distance.

Fixes: 8a5e2c9e2b16 ("obmm: Add memory region import functionality")

Signed-off-by: yuhao_zhang <yuhao.zhang@huawei.com>
(cherry picked from commit 83ae17336e75ca20f3819afd45e2067346ca6f77)
Signed-off-by: Liu Jian <liujian56@huawei.com>