Packet throughput decreases with multiple jetty streams
ANBZ: #IJSBBS

commit e3710471e4d39dbfc7862eb463b1f043d20a1e71 openEuler

1. The main modifications in this patch are as follows:

1. Symptom: When using RC write/read packet rate, with 8 ports,
the total packet rate for 32 jetties is 113M, and for 8 jetties
it is 254M.Increasing the number of jetties leads to a decrease
in the total packet rate.

2. Root Cause: The 4K page table has 2K page table entries,
and the UMMU-TLB has a capacity of 1K, causing UMMU-TLB overflow.
TLB misses occur under multiple jetty scenarios.

3. Solution: JFS's write-queue and cqe use 2M large pages.
The UMMU-TLB can accommodate a maximum of 2GB of
address access, supporting 8K depths of 2048 (JFS + JFC).

Fixes: 9b424fbe7481 ("urma: fixed process exiting uburma driver's private memory UAF")
Signed-off-by: Wen Chen <chenwen54@huawei.com>
Signed-off-by: qiushengming <qiushengming1@huawei.com>